requirements.txt
panda3d
panda3d-pman